==> WorkshopCodes/Shaastra2026-IITMadras/02-ClickStreamAnalysis/01-GenerateClickStreamData.ipynb <==
import pandas as pd
import numpy as np
import datetime as dt

# Click stream data generation 

###############################################################################    
def gensessid(n):
    '''
    function to generate random numbers
    '''
    for i in range(n):
        z = np.random.randint(5, 25)
        yield (z)
###############################################################################    
def gentime(n):
    '''
    function to generate random time events
    '''
    for i in range(n):
        z = np.random.randint(5, 345)
        yield (z)
###############################################################################
def genchannels():
    '''
    function to generate channels
    '''
    #channel
    channels = ['goog', 'fb', 'insta', 'twit']

    c1 = np.random.random()
    if (c1 < 0.4):
        c= []
    else:
        c = np.random.choice(channels, p=[0.1, 0.3, 0.5, 0.1])
    return (c)
###############################################################################
def picksession(sesslist, gens):
    '''
    function to pick session id
    '''
    z = np.random.random()
    if (z<0.7):
        ss = np.random.choice(sesslist)
    else:
        ss = sesslist[-1] + next(gens)
    return (ss)       
###############################################################################
def getproid(prlist):
    '''
    function to return product id
    '''    
    z = np.random.random()
    if (z<0.25):
        pid = prlist[0]
    elif (z< 0.45):
        pid = prlist[1]
    else:
        pid = np.random.choice(prlist[2:])
    return (pid)
###############################################################################    
def buy(pid, prlist):
    ''''
    function to return a campaign 
    '''
    if (pid==prlist[0]):
        f = int(np.random.random() + 0.6)
    elif (pid==prlist[1]):
        f = int(np.random.random() + 0.05)
    else:
        f = int(np.random.random() + 0.2)
    return (f)
###############################################################################    
def campaign(e):
    ''''
    function to return a campaign 
    '''
    if (e==1):
        f = int(np.random.random() + 0.6)
    else:
        f = 0
    return (f)
###############################################################################    
def purchasevalue(pid, prodid, prodcost):
    ''''
    function to return a purchase value 
    '''
    indx = np.where(prodid == pid)
    # add random discount
    cost = prodcost[indx] * (1. - 0.05 * np.random.random())
    return (round(float(cost), 1))

#MAIN PROGRAM
    
#session id
ses_start = 129894938431

#date
strt_date = dt.datetime.strptime("2025-11-01 09:00:00", "%Y-%m-%d %H:%M:%S")

#product ids
prodid = np.random.choice(range(1001, 1543), 60, replace=False)
# product costs
prodcost = np.random.randint(40, 240, 60)

# observations
nobs = 100000

sessgen = gensessid(nobs)
timegen = gentime(nobs)

data = []
sesslist = []
ff = 0
for i in range(nobs):
    if (ff==0):
        a = ses_start
        b = strt_date
        c = genchannels()
        d = getproid(prodid)
        e = buy(d, prodid)
        f = campaign(e)
        g = e * purchasevalue(d, prodid, prodcost)
        sesslist.append(a)
        ff= -1
    else:
        a = picksession(sesslist, sessgen)
        b += dt.timedelta(seconds = next(timegen))
        c = genchannels()
        d = getproid(prodid)
        e = buy(d, prodid)
        f = campaign(e)
        g = e * purchasevalue(d, prodid, prodcost)
    
        
        if (a not in sesslist):
            sesslist.append(a)
    
    data.append([a,b,c,d, e,f, g])
    
ddf = pd.DataFrame(data)
ddf.columns = ['session', 'date_time', 'channel', 'product_id', 'buy', 'campaign', 'value']
ddf.to_csv('clickstream.csv')


ddf.tail()

ddf.head()

ddf['buy'].value_counts()

ddf['campaign'].value_counts()

# Move date back
df = pd.read_csv('clickstream.csv', index_col=0)
df.tail()
df['ndt'] = df['date_time'].copy()
df['date_time'] = pd.to_datetime(df['ndt']) - pd.Timedelta(weeks=20)

df.tail()

df['date_time'] = df['date_time'].astype(str)
df.head()

temp = list(df.columns)
newdf = df[temp[:-1]]
newdf.head()

newdf.to_csv('clickstream.csv', index=False)

newdf.head()
