==> WorkshopCodes/Shaastra2026-IITMadras/02-ClickStreamAnalysis/03-CreateCampaignChains.ipynb <==
import pandas as pd
import numpy as np
import datetime as dt
import matplotlib.pyplot as plt
import csv

# Read the data

#df = pd.read_csv('clickstream.csv', index_col=0)
df = pd.read_csv('clickstream.csv')
purtrack = []
campdict = {}

###############################################################################
def writedict(mydict, filename):
    with open(filename, 'w') as f:
        writer = csv.writer(f)
        for key, value in mydict.items():
            writer.writerow([key, value])
    return ()

# Generate Campaign Chains

for i in range(len(df)):
    sid = df.session[i]
    time = df.date_time[i]
    chnl = df.channel[i]
    buy = df.buy[i]
    pid = df.product_id[i]
    camp = df.campaign[i]
    value = df.value[i]
    
    if (chnl != '[]'):
        if sid in campdict.keys():
            campdict[sid].append([chnl, time])
        else:
            campdict[sid] = [[chnl, time]]
            
    if(buy == 1):
        try:
            purtrack.append([campdict[sid], pid, camp, value])
            del campdict[sid]
        except:
            purtrack.append([[chnl, time], pid, camp, value])
        
purdf = pd.DataFrame(purtrack)
purdf.columns = ['chain', 'pid', 'campaign', 'value']
purdf.to_csv('Purchase_Chain.csv')

# Write non purchases to dictionary
writedict(campdict, 'NonPurchase_Chain.csv')

# Create Markov Chains

###############################################################################    
def markchain(mk):
    '''
    function to generate markov chain
    '''
    mk = eval(mk)
    
    cc = []
    
    if any(isinstance(el, list) for el in mk):
        for i in range(len(mk)):
            jj = mk[i]
            #print (jj)
            if (jj[0] != '[]'):
                cc.append(jj[0])
    cct = str(cc)
    return (cct)
###############################################################################


df = pd.read_csv('Purchase_Chain.csv')
print (df.head())

markovdict = {}
mkv_value = {}
for i in range(len(df)):
    if (np.mod(i, 5000) == 0):
        print(i, df.chain[i])
    mkc = markchain(df.chain[i])
    if (mkc in markovdict.keys()):
        markovdict[mkc] += 1
    else:
        markovdict[mkc] = 1
    if (mkc in mkv_value.keys()):
        mkv_value[mkc] += df.value[i]
    else:
        mkv_value[mkc] = df.value[i]

writedict(markovdict, 'Purchase_Markov_Count.csv')
writedict(mkv_value, 'Purchase_Markov_Value.csv')

df2 = pd.read_csv('NonPurchase_Chain.csv', header=None)
df2.columns = ['session', 'chain']
print (df2.head())

umarkovdict = {}
for i in range(len(df2)):
    if (np.mod(i, 5000) == 0):
        print(i, df2.chain[i])
    mkc = markchain(df2.chain[i])
    if (mkc in umarkovdict.keys()):
        umarkovdict[mkc] += 1
    else:
        umarkovdict[mkc] = 1

writedict(umarkovdict, 'NonPurchaseMarkov_Count.csv')

#successful purchases    
df11 = pd.read_csv('Purchase_Markov_Count.csv', header=None)
df11.columns = ['Channel', 'ConversionCount']
print (df11.head())

df12 = pd.read_csv('Purchase_Markov_Value.csv', header=None)
df12.columns = ['Channel', 'Value']
print (df12.head())

df1 = pd.merge(df11, df12, how='outer', on = ['Channel'])
print (df1.head())


#unsuccessful purchases
df2 = pd.read_csv('NonPurchaseMarkov_Count.csv', header=None)
df2.columns = ['Channel', 'NonConversionCount']
print (df2.head())

df3 = pd.merge(df1, df2, how='outer', on = ['Channel'])
#replace NaNs with zero
df3['ConversionCount'] = df3['ConversionCount'].fillna(0)
df3['Value'] = df3['Value'].fillna(0)
df3['NonConversionCount'] = df3['NonConversionCount'].fillna(0)
df3.head()

df3.to_csv('MergedChains.csv')
