==> WorkshopCodes/Shaastra2026-IITMadras/02-ClickStreamAnalysis/04-ChannelAttribution.ipynb <==
import numpy as np
import pandas as pd
from ChannelAttribution import *
import plotly.io as pio
pio.renderers.default = 'iframe'

#pip install --upgrade setuptools
#pip install Cython
#pip install ChannelAttribution
#pip install plotly

#pip install plotly==5.18.0

#Data = pd.read_csv("https://channelattribution.io/csv/Data.csv",sep=";")
Data = pd.read_csv('SampleDataChannelAttribution.csv', index_col=0)

Data.head()

DataGroup = Data.groupby('path').agg('sum').reset_index()

DataGroup.head()

# Heuristic - First Touch, Last Touch, Linear Attribution 

H=heuristic_models(DataGroup,"path","total_conversions", var_value="total_conversion_value", flg_pro=False)

H

# Markov Chain Model

M=markov_model(DataGroup, "path", "total_conversions", var_value="total_conversion_value", flg_pro=False)
M

M2 = markov_model(Data=DataGroup, var_path= "path", var_conv="total_conversions", 
             var_null = "total_null", var_value="total_conversion_value")
M2

Data.total_conversion_value.sum()

M.total_conversion_value.sum()

def plot_attrib_conv(H, M):
    
    R=pd.merge(H,M,on="channel_name",how="inner")
    R1=R[["channel_name","first_touch_conversions","last_touch_conversions",\
    "linear_touch_conversions","total_conversions"]]
    R1.columns=["channel_name","first_touch","last_touch","linear_touch","markov_model"]
    R1=pd.melt(R1, id_vars="channel_name")

    data = [dict(
          type = "histogram",
          histfunc="sum",
          x = R1.channel_name,
          y = R1.value,
          transforms = [dict(
            type = "groupby",
            groups = R1.variable,
          )],
        )]

    fig = dict({"data":data})
    pio.show(fig,validate=False)

    return()

def plot_attrib_value(H, M):
    R=pd.merge(H,M,on="channel_name",how="inner")
    R2=R[["channel_name","first_touch_value","last_touch_value",\
    "linear_touch_value","total_conversion_value"]]
    R2.columns=["channel_name","first_touch","last_touch","linear_touch","markov_model"]
    
    R2=pd.melt(R2, id_vars="channel_name")
    
    data = [dict(
      type = "histogram",
      histfunc="sum",
      x = R2.channel_name,
      y = R2.value,
      transforms = [dict(
        type = "groupby",
        groups = R2.variable,
      )],
    )]
    
    fig = dict({"data":data})
    pio.show(fig,validate=False)

    return ()

plot_attrib_conv(H, M)

plot_attrib_value(H, M)

# On Clickstream Data

# On our data
DF = pd.read_csv('MergedChains.csv', index_col= 0)
DF.head()

def cast_list(ll):
    '''
    function to convert clickstream string to required format with > as separator'
    '''
    nll = eval(ll)
    strr = ''
    for x in nll:
        if strr == '':
            strr += x
        else:
            strr += ' > ' + x
    return (strr)

# Test function 
ll = str(['fb', 'fb', 'fb', 'fb', 'fb'])
cast_list(ll)

DF['newchannel'] = DF['Channel'].apply(lambda x: cast_list(x))
DF1 = DF[['newchannel', 'ConversionCount', 'Value', 'NonConversionCount']]
DF1.columns = ['path', 'total_conversions', 'total_conversion_value', 'total_null']
DF1.to_csv('TestChains.csv')
DF1.head()

M=markov_model(DF1, "path", var_conv = "total_conversions", 
               var_null = "total_null", var_value="total_conversion_value", flg_pro=False)
M

H=heuristic_models(DF1,"path", "total_conversions",var_value="total_conversion_value", flg_pro=False)
H

DF1.total_conversion_value.sum()

M.total_conversion_value.sum()

DF1.total_conversions.sum()

M.total_conversions.sum()

plot_attrib_conv(H, M)

plot_attrib_value(H, M)
